requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
nonebot2>=2.0.0
nonebot-plugin-htmlrender>=0.2.0
nonebot-plugin-alconna>=0.4.0
nonebot-adapter-onebot>=2.0.0